+++ Sources/MessagingInApp/Gist/Managers/MessageManager.swift
@@ -16,7 +16,7 @@ public enum GistMessageActions: String {
  * Override any of the abstract functions in class to implement custom logic for when certain events happen. Depending on the type of message you are displaying, you may want to handle events differently.
  */
 class MessageManager {
-    private var engine: EngineWebInstance
+    var engine: EngineWebInstance
     private let siteId: String
     let currentMessage: Message
     var gistView: GistView!
@@ -28,21 +28,21 @@ class MessageManager {
     init(siteId: String, message: Message) {
         self.siteId = siteId
         self.currentMessage = message
-        self.currentRoute = message.messageId
+        self.currentRoute = message.templateId
 
         let engineWebConfiguration = EngineWebConfiguration(
             siteId: Gist.shared.siteId,
             dataCenter: Gist.shared.dataCenter,
             instanceId: message.instanceId,
             endpoint: Settings.Network.engineAPI,
-            messageId: message.messageId,
+            messageId: message.templateId,
             properties: message.toEngineRoute().properties
         )
 
         // When EngineWeb instance is constructed, it will begin the rendering process for the in-app message.
         // This means that the message begins the process of loading.
         // Start a timer that helps us determine how long a message took to load/render.
-        elapsedTimer.start(title: "Loading message with id: \(currentMessage.messageId)")
+        elapsedTimer.start(title: "Loading message with id: \(currentMessage.templateId)")
 
         self.engine = engineWebProvider.getEngineWebInstance(configuration: engineWebConfiguration)
         engine.delegate = self
@@ -83,7 +83,7 @@ extension MessageManager: EngineWebDelegate {
         Logger.instance.debug(message: "Bourbon Engine bootstrapped")
 
         // Cleaning after engine web is bootstrapped and all assets downloaded.
-        if currentMessage.messageId == "" {
+        if currentMessage.templateId == "" {
             engine.cleanEngineWeb()
         }
     }
@@ -192,7 +192,7 @@ extension MessageManager: EngineWebDelegate {
     }
 
     func error() {
-        Logger.instance.error(message: "Error loading message with id: \(currentMessage.messageId)")
+        Logger.instance.error(message: "Error loading message with id: \(currentMessage.templateId)")
         delegate?.messageError(message: currentMessage)
     }