+++ Tests/MessagingInApp/Extensions/GistExtensions.swift
@@ -2,13 +2,16 @@
 import Foundation
 
 extension Message {
-    convenience init(messageId: String = .random, campaignId: String = .random, queueId: String = .random, elementId: String? = nil, priority: Int? = nil) {
+    convenience init(messageId: String = .random, campaignId: String = .random, pageRule: String? = nil, queueId: String = .random, elementId: String? = nil, priority: Int? = nil) {
         var gistProperties = [
             "campaignId": campaignId
         ]
 
         if let elementId = elementId {
             gistProperties["elementId"] = elementId
+        }
+        if let pageRule = pageRule {
+            gistProperties["routeRuleApple"] = pageRule
         }
 
         self.init(queueId: queueId, priority: priority, messageId: messageId, properties: [